JENA-1594: Structure to allow the adding/modifying of server setup.
